+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This file is used to list changes made in each version of the AWS ParallelCluste
 
 **CHANGES**
 - Ubuntu 16.04 is no longer supported.
+- Amazon Linux is no longer supported.
 
 
 2.10.1